refactor(decomp): hoist CollectionIdentity + CollectionPathResolver port to crates (Slice D links 3-4)
Two root-crate-decomposition hoists that free `src/storage`/`src/core` of two
root-internal dependencies via port-inversion, shrinking the monolithic root
crate (the OOM/compile-time pressure driver). Both are clean-leaf moves with
re-export shims left in root so all existing `crate::...` callers compile
unchanged — no call-site edits.

Part A — CollectionIdentity (clean leaf) → foundation kernel
- `src/core/stable_id.rs` (192 lines: `CollectionIdentity` + `AccountId`/
  `NamespaceId`/`CollectionId`/`ColumnId`/`IndexId`/`SegmentId`/`WorkspaceId`
  type aliases + `ToPathSegment` zero-padded base62 trait) moved verbatim to
  `crates/foundation/proximadb-kernel/src/stable_id.rs`. Its only dependency was
  `proximadb_kernel::base62`, which is in the same crate — confirmed zero
  `use crate::...` references before the move.
- Kernel `lib.rs`: added `pub mod stable_id;` + `pub use stable_id::{...}`.
- Root `src/core/stable_id.rs` is now a 14-line re-export shim
  (`pub use proximadb_kernel::stable_id::{...};`).
- 5 tests moved to the kernel crate (all pass).

Part B — CollectionPathResolver port → storage-ports crate
- The `CollectionPathResolver` trait + `StorageAssignment` value type
  (primitive-typed signatures: `String`/`bool`/`u32`/`Vec<String>` only —
  confirmed the trait body references NO root-internal type) moved to
  `crates/storage/proximadb-storage-ports/src/path_resolver.rs`.
- storage-ports `lib.rs`: added `pub mod path_resolver;` + `pub use`.
- Root `src/storage/trait_components/path_resolver.rs` now re-exports the port
  (`pub use proximadb_storage_ports::{CollectionPathResolver, StorageAssignment};`)
  and KEEPS the root-catalog-coupled concrete impls (`ConfigFallbackResolver`,
  `CachedResolver`, `CompositeResolver`) + `DrResolvedPath`/`DrPathBuilder` —
  these `impl` the crate trait (root→storage-ports is a layering-allowed
  downward edge; the forbidden direction was storage→root-internal).
- `test_storage_assignment_default` moved to the crate (1 test, passes); the
  impl tests stay in root (they test the root-resident impls).

Verification (all green):
- `cargo check --lib --bins`, `cargo check --tests -p proximadb`,
  `-p proximadb-kernel`, `-p proximadb-storage-ports`.
- `cargo nextest run -p proximadb-kernel` (143 passed incl. 5 moved) +
  `-p proximadb-storage-ports` (1 moved test passed).
- `cargo clippy --lib --bins` (exit 0); `cargo fmt --all && cargo fmt --check`.
- `make work-commit-check` (fmt-check, docs-claim, capability-matrix,
  workspace-boundaries, tenant-path, panic-policy module guard, hygiene) — all
  passed; panic-policy delta -1 (the root shim dropped an item).

Part of the Slice-D root-crate-decomposition chain (links 1-2: BatchId +
BlockPruneMode in #896). No on-disk/wire format change — pure code move.
