+++ crates/foundation/proximadb-kernel/src/lib.rs
@@ -17,6 +17,7 @@ pub mod metadata_error;
 pub mod request_limits;
 pub mod score_types;
 pub mod service_error;
+pub mod stable_id;
 pub mod stream_error;
 pub mod uuid;
 
@@ -33,5 +34,9 @@ pub use metadata_error::MetadataError;
 pub use request_limits::RequestLimits;
 pub use score_types::{PhaseId, ScoreComponent, ScoreVector};
 pub use service_error::ServiceError;
+pub use stable_id::{
+    AccountId, CollectionId, CollectionIdentity, ColumnId, IndexId, NamespaceId, SegmentId,
+    ToPathSegment, WorkspaceId,
+};
 pub use stream_error::{StreamError, StreamResult};
 pub use uuid::*;
