+++ src/storage/trait_components/path_resolver.rs
@@ -50,84 +50,13 @@ use proximadb_catalog::{
 use proximadb_storage_common::StoragePath;
 use std::sync::Arc;
 
-/// Storage location assignment for a collection
-#[derive(Debug, Clone)]
-pub struct StorageAssignment {
-    /// Primary storage URL (e.g., "file:///data/proximadb/d1")
-    pub primary_url: String,
-    /// Weight for load balancing (1-100)
-    pub weight: u32,
-    /// Whether this location is available
-    pub available: bool,
-    /// Optional: Replica URLs for high availability
-    pub replica_urls: Vec<String>,
-}
-
-impl Default for StorageAssignment {
-    fn default() -> Self {
-        Self {
-            primary_url: "file:///tmp/proximadb/data".to_string(),
-            weight: 1,
-            available: true,
-            replica_urls: Vec::new(),
-        }
-    }
-}
-
-/// Collection path resolver trait (DIP-compliant interface)
-///
-/// Abstracts the resolution of storage paths for collections,
-/// replacing global singletons with dependency injection.
-#[async_trait]
-pub trait CollectionPathResolver: Send + Sync {
-    /// Resolver name for logging/debugging
-    fn name(&self) -> &'static str;
-
-    /// Resolve the base storage location for a collection
-    ///
-    /// # Arguments
-    /// * `collection_id` - The collection identifier
-    ///
-    /// # Returns
-    /// The base URL for the collection's storage (e.g., "file:///data/proximadb/collections/my_collection")
-    async fn resolve_base_location(&self, collection_id: &str) -> Result<String>;
-
-    /// Resolve the storage assignment for a collection
-    ///
-    /// # Arguments
-    /// * `collection_id` - The collection identifier
-    ///
-    /// # Returns
-    /// Storage assignment details including primary URL and replicas
-    async fn resolve_storage_assignment(&self, collection_id: &str) -> Result<StorageAssignment>;
-
-    /// Resolve the WAL directory for a collection
-    ///
-    /// # Arguments
-    /// * `collection_id` - The collection identifier
-    ///
-    /// # Returns
-    /// The WAL directory URL (e.g., "file:///data/proximadb/collections/my_collection/wal")
-    async fn resolve_wal_location(&self, collection_id: &str) -> Result<String> {
-        let base = self.resolve_base_location(collection_id).await?;
-        Ok(format!("{}/wal", base))
-    }
-
-    /// Resolve the SST directory for a collection
-    ///
-    /// # Arguments
-    /// * `collection_id` - The collection identifier
-    ///
-    /// # Returns
-    /// The SST files directory URL
-    async fn resolve_sst_location(&self, collection_id: &str) -> Result<String> {
-        let base = self.resolve_base_location(collection_id).await?;
-        Ok(format!("{}/sst", base))
-    }
-
-    /// Check if a collection exists
-    async fn collection_exists(&self, collection_id: &str) -> Result<bool>;
-}
+// ─── Port re-exports (Slice D hoist) ───────────────────────────────────────
+// The `CollectionPathResolver` trait + `StorageAssignment` value type now live
+// in `proximadb_storage_ports` (a clean, facade-free port — primitive-typed
+// signatures only). The root-catalog-coupled concrete impls below stay here
+// and `impl` the crate trait. Re-exported so existing callers
+// (`crate::storage::trait_components::path_resolver::*`) keep compiling.
+pub use proximadb_storage_ports::{CollectionPathResolver, StorageAssignment};
 
 // ============================================================================
 // Standard Implementations
